Fixed GUI, cache & search related bugs, excluding a directory from caching

+++ src/net/quickfind/cache/CachePage.java
@@ -45,6 +45,7 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import java.util.zip.ZipEntry;
 import java.util.zip.ZipInputStream;
+import javax.swing.JOptionPane;
 import javax.xml.bind.JAXBContext;
 import javax.xml.bind.JAXBException;
 import javax.xml.bind.Marshaller;
@@ -79,11 +80,26 @@ public class CachePage {
      * Loads the cache configuration from file
      */
     public void loadCacheCollection() {
-        if (!loadCacheConfigFromFile(PropertyPage.CACHE_COLLECTION_CONFIG_FILE + PropertyPage.CONFIG_FILE_EXTENSION)) {
+        boolean isCacheConfigLoadSuccess = loadCacheConfigFromFile(PropertyPage.CACHE_COLLECTION_CONFIG_FILE + PropertyPage.CONFIG_FILE_EXTENSION);
+        if (!isCacheConfigLoadSuccess || cacheCollection.getCachesList() == null) {
+            Utility.cacheCleaner(null, 1);
             cacheCollection = new CacheCollection();
             cacheCollection.setCacheList(new ArrayList<Cache>());
             cacheCollection.setName(PropertyPage.getSystemName());
             cacheCollection.setCacheIndex("-1");
+        }
+        reLoadCacheList();
+    }
+
+    /*
+     * Reloads the caches from cache collection
+     */
+    public void reLoadCacheList() {
+        PropertyPage.CACHE_LIST.clear();
+        if (cacheCollection.getCachesList() != null) {
+            for (int i = 0; i < cacheCollection.getCachesList().size(); i++) {
+                PropertyPage.CACHE_LIST.add(((Cache) cacheCollection.getCachesList().get(i)).getCacheName());
+            }
         }
     }
 
@@ -91,8 +107,10 @@ public class CachePage {
      * Exctracts raw files from compressed file
      */
     public void generateRawData() {
-        for (int i = 0; i < cacheCollection.getCachesList().size(); i++) {
-            createRawData(((Cache) cacheCollection.getCachesList().get(i)).getName());
+        if (cacheCollection.getCachesList() != null) {
+            for (int i = 0; i < cacheCollection.getCachesList().size(); i++) {
+                createRawData(((Cache) cacheCollection.getCachesList().get(i)).getCacheName());
+            }
         }
     }
 
@@ -108,14 +126,44 @@ public class CachePage {
      * @param   cache object which is to be inserted
      */
     protected void addCache(Cache cache) {
+        if (cache == null) {
+            return;
+        }
         for (int i = 0; i < cacheCollection.getCachesList().size(); i++) {
-            if (((Cache) cacheCollection.getCachesList().get(i)).getName().equals(cache.getName())) {
+            if (((Cache) cacheCollection.getCachesList().get(i)).getCacheName().equals(cache.getCacheName())) {
                 cacheCollection.addCache(cache, i);//add new cache
                 cacheCollection.getCachesList().remove(i + 1);//remove old cache
                 return;
             }
         }
         cacheCollection.addCache(cache, cacheCollection.getCachesList().size());   //add new cache
+    }
+
+    /*
+     * Updates cache list with new cache or replaces cahe if it exists already
+     * @param    cache object which is to be updated
+     */
+    public void updateCache(Cache cache) {
+        addCache(cache);
+        saveCacheConfigToFile();
+    }
+
+    /*
+     * Removes cache permanently
+     * @param   User given name of the cache which is to be deleted
+     */
+    public void removeCache(String userGivenName) {
+        deleteOldCacheFiles(getCacheName(userGivenName));
+        ArrayList<Cache> cacheList = cacheCollection.getCachesList();
+        if (cacheList != null) {
+            for (int i = 0; i < cacheList.size(); i++) {
+                if (((Cache) cacheList.get(i)).getuserGivenName().equals(userGivenName)) {
+                    cacheList.remove(i);
+                }
+            }
+        }
+        PropertyPage.CACHE_LIST.remove(getCacheName(userGivenName));
+        saveCacheConfigToFile();
     }
 
     /*
@@ -170,7 +218,17 @@ public class CachePage {
                 JAXBContext jAXBContext = JAXBContext.newInstance(CacheCollection.class);
                 Unmarshaller unMarshaller = jAXBContext.createUnmarshaller();
                 cacheCollection = (CacheCollection) unMarshaller.unmarshal(new FileReader(cacheConfigFile));
-                return true;
+                if (evaluateConfig(cacheCollection)) {
+                    return true;
+                } else {
+                    int status = JOptionPane.showConfirmDialog(null, "It looks like your cache configuration is invalid/old!\nQF wants to clean it up. Do you want to proceed?", "Attention", JOptionPane.YES_NO_OPTION);
+                    if (status == JOptionPane.YES_OPTION) {
+                        return false;
+                    } else if (status == JOptionPane.NO_OPTION) {
+                        System.out.println("QuickFind Exiting...");
+                        System.exit(0);
+                    }
+                }
             } catch (JAXBException jAXBException) {
                 LOGGER.log(Level.SEVERE, "An error was encountered while creating the JAXBContext", jAXBException);
             } catch (IOException iOException) {
@@ -181,6 +239,28 @@ public class CachePage {
     }
 
     /*
+     * Evaluates mandatory cache properties
+     * @param cache collection to be evaluated
+     */
+    private boolean evaluateConfig(CacheCollection cacheCollection) {
+        if (cacheCollection != null) {
+            ArrayList<Cache> cacheList = cacheCollection.getCachesList();
+            if (cacheList != null) {
+                for (int i = 0; i < cacheList.size(); i++) {
+                    Cache cache = (Cache) cacheList.get(i);
+                    if (cache.getCacheName() == null || cache.getCacheName().isEmpty()
+                            || cache.getuserGivenName() == null || cache.getuserGivenName().isEmpty()
+                            || cache.getIncludedPath() == null || cache.getIncludedPath().isEmpty()
+                            || cache.getLocalCacheIndex() == null || cache.getLocalCacheIndex().isEmpty()) {
+                        return false;
+                    }
+                }
+            }
+        }
+        return true;
+    }
+
+    /*
      * Exctracts raw files from compressed file
      * @param   compressed cahe file name
      */
@@ -188,23 +268,26 @@ public class CachePage {
         byte[] buffer = new byte[1024];
         int len;
         try {
-            ZipInputStream zipInputStream = new ZipInputStream(new FileInputStream(PropertyPage.getCacheDirectory() + zipFileName + PropertyPage.COMPRESSED_FILE_EXTENSION));
-            //get the zipped file entry
-            ZipEntry zipEntry = zipInputStream.getNextEntry();
-            while (zipEntry != null) {
-                String fileName = zipEntry.getName();
-                File newFile = new File(PropertyPage.getCacheDirectory() + PropertyPage.FILE_SEPARATOR + fileName);
-                //creates all non exists directories
-                new File(newFile.getParent()).mkdirs();
-                FileOutputStream fileOutputStream = new FileOutputStream(newFile);
-                while ((len = zipInputStream.read(buffer)) > 0) {
-                    fileOutputStream.write(buffer, 0, len);
+            File zipRawFile = new File(PropertyPage.getCacheDirectory() + zipFileName + PropertyPage.COMPRESSED_FILE_EXTENSION);
+            if (zipRawFile.exists()) {
+                ZipInputStream zipInputStream = new ZipInputStream(new FileInputStream(zipRawFile));
+                //get the zipped file entry
+                ZipEntry zipEntry = zipInputStream.getNextEntry();
+                while (zipEntry != null) {
+                    String fileName = zipEntry.getName();
+                    File newFile = new File(PropertyPage.getCacheDirectory() + PropertyPage.FILE_SEPARATOR + fileName);
+                    //creates all non exists directories
+                    new File(newFile.getParent()).mkdirs();
+                    FileOutputStream fileOutputStream = new FileOutputStream(newFile);
+                    while ((len = zipInputStream.read(buffer)) > 0) {
+                        fileOutputStream.write(buffer, 0, len);
+                    }
+                    fileOutputStream.close();
+                    zipEntry = zipInputStream.getNextEntry();
                 }
-                fileOutputStream.close();
-                zipEntry = zipInputStream.getNextEntry();
+                zipInputStream.closeEntry();
+                zipInputStream.close();
             }
-            zipInputStream.closeEntry();
-            zipInputStream.close();
         } catch (IOException iOException) {
             LOGGER.log(Level.SEVERE, "An IO error occured", iOException);
         }
@@ -214,8 +297,10 @@ public class CachePage {
      * Deletes existing cache files
      * @param   encoded name of the cache
      */
-    protected void deleteOldCacheFiles(final String cacheName) {
-        Utility.cacheCleaner(cacheName, 3);
+    public void deleteOldCacheFiles(final String cacheName) {
+        if (cacheName != null) {
+            Utility.cacheCleaner(cacheName, 3);
+        }
     }
 
     /*
@@ -225,7 +310,7 @@ public class CachePage {
      */
     private int getCacheIndex(String cacheName) {
         for (int i = 0; i < cacheCollection.getCachesList().size(); i++) {
-            if (((Cache) cacheCollection.getCachesList().get(i)).getName().equals(cacheName)) {
+            if (((Cache) cacheCollection.getCachesList().get(i)).getCacheName().equals(cacheName)) {
                 return Integer.valueOf(((Cache) cacheCollection.getCachesList().get(i)).getLocalCacheIndex());
             }
         }
@@ -245,5 +330,90 @@ public class CachePage {
         cacheIndex = Integer.valueOf(cacheCollection.getCacheIndex());
         cacheCollection.setCacheIndex(String.valueOf(++cacheIndex));
         return cacheIndex;
+    }
+
+    /*
+     * Searches cache for the given user provided cache name
+     * @param User provided name of the cache
+     * @return Cache object if it is available otherwise null
+     */
+    public Cache getCache(String userGivenCacheName) {
+        ArrayList<Cache> cacheList = cacheCollection.getCachesList();
+        if (cacheList != null) {
+            for (int i = 0; i < cacheList.size(); i++) {
+                if (((Cache) cacheList.get(i)).getuserGivenName().equals(userGivenCacheName)) {
+                    return ((Cache) cacheList.get(i));
+                }
+            }
+        }
+        return null;
+    }
+
+    /*
+     * Searches cache name for the user given name
+     * @param User given name of the cache
+     * @return Cache name if it is available otherwise null
+     */
+    public String getCacheName(String userGivenName) {
+        ArrayList<Cache> cacheList = cacheCollection.getCachesList();
+        if (cacheList != null) {
+            for (int i = 0; i < cacheList.size(); i++) {
+                if (((Cache) cacheList.get(i)).getuserGivenName().equals(userGivenName)) {
+                    return ((Cache) cacheList.get(i)).getCacheName();
+                }
+            }
+        }
+        return null;
+    }
+
+    /*
+     * Searches user given name for the given cache name
+     * @param Encoded name of the cache
+     * @return User given cache name if it is available otherwise null
+     */
+    public String getUserGivenName(String cacheName) {
+        ArrayList<Cache> cacheList = cacheCollection.getCachesList();
+        if (cacheList != null) {
+            for (int i = 0; i < cacheList.size(); i++) {
+                if (((Cache) cacheList.get(i)).getCacheName().equals(cacheName)) {
+                    return ((Cache) cacheList.get(i)).getuserGivenName();
+                }
+            }
+        }
+        return null;
+    }
+
+    /*
+     * Searches included cache path for the given cache name
+     * @param Encoded name of the cache
+     * @return Included cache path if it is available otherwise null
+     */
+    public String getIncludedCachePath(String cacheName) {
+        ArrayList<Cache> cacheList = cacheCollection.getCachesList();
+        if (cacheList != null) {
+            for (int i = 0; i < cacheList.size(); i++) {
+                if (((Cache) cacheList.get(i)).getCacheName().equals(cacheName)) {
+                    return ((Cache) cacheList.get(i)).getIncludedPath();
+                }
+            }
+        }
+        return null;
+    }
+
+    /*
+     * Checks whether cache exists for the given include & exclude paths
+     * @param includes-> Paths to be included in caching   excludes-> Paths to be excluded from caching
+     * @return true if cache found otherwise false
+     */
+    public boolean isCacheExists(String includes, String excludes) {
+        ArrayList<Cache> cacheList = cacheCollection.getCachesList();
+        if (cacheList != null) {
+            for (int i = 0; i < cacheList.size(); i++) {
+                if ((((Cache) cacheList.get(i)).getIncludedPath().equalsIgnoreCase(includes)) && (((Cache) cacheList.get(i)).getExcludedPath().equalsIgnoreCase(excludes))) {
+                    return true;
+                }
+            }
+        }
+        return false;
     }
 }
